refactor(plugin): share tracked registration teardown
