+++ crates/core/src/plugin/dynamic/native.rs
@@ -51,13 +51,13 @@ use crate::api::scope::{event as emit_scope_mark, get_handle, pop_scope, push_sc
 use crate::api::tool::ToolExecutionInterceptOutcome;
 use crate::error::{FlowError, Result as FlowResult};
 use crate::plugin::{
-    ConfigDiagnostic, DiagnosticLevel, Plugin, PluginDeregistrationOutcome, PluginError,
-    PluginRegistrationContext, deregister_plugin_registration_checked, register_plugin_tracked,
+    ConfigDiagnostic, DiagnosticLevel, Plugin, PluginError, PluginRegistrationContext,
+    deregister_plugin_registration_checked, register_plugin_tracked,
 };
 
 use super::{
     DynamicPluginKind, DynamicPluginManifest, DynamicPluginManifestLoad,
-    DynamicPluginTeardownOutcome,
+    DynamicPluginTeardownOutcome, deregister_tracked_registrations_checked,
 };
 
 /// Native plugin load request derived from host dynamic-plugin state.
@@ -89,30 +89,7 @@ impl NativePluginActivation {
     pub fn clear(self) {}
 
     pub(crate) fn deregister_plugin_kinds_checked(&mut self) -> DynamicPluginTeardownOutcome {
-        let mut outcome = DynamicPluginTeardownOutcome::success();
-        let plugin_registrations = std::mem::take(&mut self.plugin_registrations);
-        for (plugin_kind, registration_id) in plugin_registrations.into_iter().rev() {
-            match deregister_plugin_registration_checked(&plugin_kind, registration_id) {
-                Ok(PluginDeregistrationOutcome::Removed) => {}
-                Ok(PluginDeregistrationOutcome::Missing) => outcome.record_error(
-                    format!(
-                        "native plugin kind '{plugin_kind}' was not registered during teardown"
-                    ),
-                    true,
-                ),
-                Ok(PluginDeregistrationOutcome::Replaced) => outcome.record_error(
-                    format!(
-                        "native plugin kind '{plugin_kind}' was replaced during teardown and was left registered"
-                    ),
-                    true,
-                ),
-                Err(error) => outcome.record_error(
-                    format!("failed to deregister native plugin kind '{plugin_kind}': {error}"),
-                    false,
-                ),
-            }
-        }
-        outcome
+        deregister_tracked_registrations_checked(&mut self.plugin_registrations, "native")
     }
 
     #[cfg(test)]
